requirements.txt
litestar
uvicorn[standard]
msgspec
sniffio
python-arango
arq
redis
cryptography
google-ads
